requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile setup.cfg
#
pydantic==1.10.4
    # via builderer (setup.cfg)
pyyaml==6.0
    # via builderer (setup.cfg)
typing-extensions==4.4.0
    # via pydantic
